可以加载NPCs在走动时。但是更多NPCs Model 不被加载。并跟随走动会提示错误。Debug Version

+++ ArcheAge/ArcheAge/Structuring/NPC/NPC.cs
@@ -37,7 +37,7 @@ namespace ArcheAgeGame.ArcheAge.Structuring.NPC
 		/// <summary>
 		/// scale 中文不知道
 		/// </summary>
-		public UInt32 Scale { get; set; }
+		public float Scale { get; set; }
 
 		/// <summary>
 		/// Level
